Complete the connectivity of all the parts for the http shell.

+++ src/core/logmanager.cpp
@@ -50,7 +50,7 @@ LogManager::LogManager()
     std::string fileRel("src/core/logmanager.cpp");
     auto p = filePP.find(fileRel);
     if (p != std::string::npos) {
-        srcFilePrefixLength = p;
+        srcFilePrefixLength = (int)p;
     }
 }
